test(content-system): split codec and descriptor suites along the wiring extraction
